Test common

+++ interop/testlib/classes/subnetmanager.py
@@ -25,28 +25,37 @@ class SubnetManager:
     def start(self):
         """ Starts the subnet manager at ip address
         """
-        self.node.command("systemctl start opensm")
+        if self.node.ethif.state == "up":
+            self.node.command("systemctl start opensm")
+            return True
+        else:
+            return False
 
     def stop(self):
         """ Stop the subnet manager at ip address
         """
-        self.node.command("systemctl stop opensm")
+        if self.node.ethif.state == "up":
+            self.node.command("systemctl stop opensm")
+            return True
+        else:
+            return False
 
     def status(self):
         """ Status of the subnet manager at the ip address
         """
-
-        # opensm for status
-        output = self.node.command("systemctl status opensm")
-        active_lines = []
-        for line in output.split('\n'):
-            if "Active: " in line:
-                active_lines.append(line)
-        if len(active_lines) != 1:
-            raise SubnetManagerParsingError("the output of `systemctl status opensm` was parsed incorrectly")
-        else:
-            self.state = active_lines[0].strip().split()[1]
-            return self.state
+        if self.node.ethif.state == "up":
+            # opensm for status
+            output = self.node.command("systemctl status opensm")
+            active_lines = []
+            for line in output.split('\n'):
+                if "Active: " in line:
+                    active_lines.append(line)
+            if len(active_lines) != 1:
+                raise SubnetManagerParsingError("the output of `systemctl status opensm` was parsed incorrectly")
+            else:
+                self.state = active_lines[0].strip().split()[1]
+                return self.state
+        return False
 
     def print(self):
         """ Print the the subnet manager status
